:recycle: Split App into multiple files
