+++ modules/cli/src/main/kotlin/datamaintain/cli/utils/LogUtils.kt
@@ -1,0 +1,5 @@
+package datamaintain.cli.utils
+
+import mu.KotlinLogging
+
+val logger = KotlinLogging.logger {}
